requirements.txt
Flask==3.0.0
gunicorn==21.2.0
agno==0.1.0
markdown2==2.4.10
python-dotenv==1.0.0
Pillow==10.1.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
pandas==2.1.4
plotly==5.18.0
python-multipart==0.0.6